backend/requirements.txt
websockets>=12.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=1.0.0
httpx>=0.27.0
